refactor: 위치가 변경된 button, logo 다시 import
